third_party/WebKit/LayoutTests/webaudio/realtimeanalyser-freq-data-expected.txt

Test AnalyserNode getFloatFrequencyData and getByteFrequencyData, no Smoothing

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS 32-point float FFT equals [-38.0,-14.8,-15.6,-19.9,-22.9,-25.1,-26.7,-28.1,-29.3,-30.4,-31.7,-33.7,-36.0,-38.5,-43.7,-55.3] with an element-wise tolerance of {absoluteThreshold: 0, relativeThreshold: 6.8964e-7}.
PASS Min FFT value is less than or equal to -50.
PASS Max FFT value is greater than or equal to -30.
PASS 32-point byte FFT equals [153,255,255,255,255,255,255,255,255,249,233,207,178,146,79,0] with an element-wise tolerance of 0.
PASS 64-point float FFT equals [-15.9,-17.2,-23.0,-30.0,-32.4,-34.3,-35.9,-37.3,-38.4,-39.5,-40.4,-41.2,-41.9,-42.7,-43.3,-43.9,...] with an element-wise tolerance of {absoluteThreshold: 0, relativeThreshold: 0.0000068366}.
PASS Min FFT value is less than or equal to -100.
PASS Max FFT value is greater than or equal to -30.
PASS 64-point byte FFT equals [255,255,255,254,246,239,233,228,224,220,217,214,211,208,206,204,...] with an element-wise tolerance of 0.
PASS 128-point float FFT equals [-29.5,-21.3,-20.2,-27.9,-29.0,-31.4,-34.1,-33.5,-37.0,-36.2,-37.6,-39.2,-38.3,-41.4,-39.7,-41.7,...] with an element-wise tolerance of {absoluteThreshold: 0, relativeThreshold: 0.0000014602}.
PASS Min FFT value is less than or equal to -100.
PASS Max FFT value is greater than or equal to -30.
PASS 128-point byte FFT equals [255,255,255,255,255,249,239,242,229,232,227,221,224,213,219,212,...] with an element-wise tolerance of 0.
PASS 256-point float FFT equals [-72.0,-53.6,-28.6,-19.7,-20.3,-29.9,-28.3,-25.0,-30.6,-35.8,-28.9,-30.5,-39.9,-33.4,-31.2,-37.9,...] with an element-wise tolerance of {absoluteThreshold: 0, relativeThreshold: 8.4828e-7}.
PASS Min FFT value is less than or equal to -100.
PASS Max FFT value is greater than or equal to -30.
PASS 256-point byte FFT equals [102,169,255,255,255,255,255,255,252,233,255,253,218,242,250,226,...] with an element-wise tolerance of 0.
PASS 512-point float FFT equals [-99.7,-95.6,-90.0,-86.2,-82.1,-36.4,-22.3,-19.0,-24.7,-42.6,-83.7,-79.3,-39.9,-27.4,-25.2,-32.1,...] with an element-wise tolerance of {absoluteThreshold: 0, relativeThreshold: 0.000016568}.
PASS Min FFT value is less than or equal to -100.
PASS Max FFT value is greater than or equal to -30.
PASS 512-point byte FFT equals [0,16,36,50,65,231,255,255,255,208,59,75,219,255,255,247,...] with an element-wise tolerance of 0.
PASS 1024-point float FFT equals [-100,-100,-100,-100,-99.4,-96.6,-93.6,-90.2,-86.4,-82.5,-79.7,-70.4,-33.9,-21.4,-19.2,-26.1,...] with an element-wise tolerance of {absoluteThreshold: 0, relativeThreshold: 0.000012765}.
PASS Min FFT value is less than or equal to -100.
PASS Max FFT value is greater than or equal to -30.
PASS 1024-point byte FFT equals [0,0,0,0,2,12,23,35,49,63,74,107,240,255,255,255,...] with an element-wise tolerance of 0.
PASS 2048-point float FFT equals [-100,-100,-100,-100,-100,-100,-100,-100,-100,-100,-100,-100,-100,-100,-100,-100,...] with an element-wise tolerance of {absoluteThreshold: 0, relativeThreshold: 0.000008492}.
PASS Min FFT value is less than or equal to -100.
PASS Max FFT value is greater than or equal to -30.
PASS 2048-point byte FFT equals [0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,...] with an element-wise tolerance of 0.
PASS 4096-point float FFT equals [-100,-100,-100,-100,-100,-100,-100,-100,-100,-100,-100,-100,-100,-100,-100,-100,...] with an element-wise tolerance of {absoluteThreshold: 0, relativeThreshold: 4.3344e-7}.
PASS Min FFT value is less than or equal to -100.
PASS Max FFT value is greater than or equal to -30.
PASS 4096-point byte FFT equals [0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,...] with an element-wise tolerance of 0.
PASS 8192-point float FFT equals [-100,-100,-100,-100,-100,-100,-100,-100,-100,-100,-100,-100,-100,-100,-100,-100,...] with an element-wise tolerance of {absoluteThreshold: 0, relativeThreshold: 2.2798e-7}.
PASS Min FFT value is less than or equal to -100.
PASS Max FFT value is greater than or equal to -30.
PASS 8192-point byte FFT equals [0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,...] with an element-wise tolerance of 0.
PASS 16384-point float FFT equals [-100,-100,-100,-100,-100,-100,-100,-100,-100,-100,-100,-100,-100,-100,-100,-100,...] with an element-wise tolerance of {absoluteThreshold: 0, relativeThreshold: 1.1756e-7}.
PASS Min FFT value is less than or equal to -100.
PASS Max FFT value is greater than or equal to -30.
PASS 16384-point byte FFT equals [0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,...] with an element-wise tolerance of 0.
PASS 32768-point float FFT equals [-100,-100,-100,-100,-100,-100,-100,-100,-100,-100,-100,-100,-100,-100,-100,-100,...] with an element-wise tolerance of {absoluteThreshold: 0, relativeThreshold: 1.1675e-7}.
PASS Min FFT value is less than or equal to -100.
PASS Max FFT value is greater than or equal to -30.
PASS 32768-point byte FFT equals [0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,...] with an element-wise tolerance of 0.
PASS Basic frequency data computed correctly.

PASS 128-point float FFT equals [-33.7,-21.1,-18.0,-20.2,-23.2,-25.3,-27.0,-28.4,-29.6,-30.7,-31.6,-32.4,-33.2,-33.9,-34.5,-35.2,...] with an element-wise tolerance of {absoluteThreshold: 0, relativeThreshold: 0.0000012548}.
PASS Smoothing constant of 0 correctly handled.

PASS successfullyParsed is true

TEST COMPLETE
